main: split job description code into its own file
